Oops it was async
In my previous commit I removed a setTimeout which waited 1.5 seconds before showing the contacts. It seemed to work, I guess because the contacts were already loaded, but I didn't realize it might be asynchronous otherwise. This should fix that case, without having to wait 1.5 seconds if the contacts are already loaded.