Removed unnecessary savings of messages and conversation
When the user opens a conversation all messages are loaded from the storage and added to the message collection of the conversation. Each added message fires the 'add'-event of the message collection. The conversation's `_onAddMessage` handler gets triggered, and saves the message to its storage again, which is unnecessary. Additionaly it refreshes the *last message* information of the conversation and saves the conversation to the storage.

So each loaded message in a conversation triggers a save process for the message and the conversation.

This commit addresses this issue and saves message and conversation only if there are new messages.

Another small improvement is the reuse of the `updateLastMessage` instead of a similar implementation in the `_onAddMessage` itself.