Dead (maybe for now?)
Thanks everyone for all the things you did for supporting OpenWapp.

Thanks to the guys that started the project, and let me open source it under the
Mozilla Hispano umbrella.

It has been a fun ride, with a lot of hiccups, but that worth it for the future of
Firefox OS.

Thanks the developers, from Salva to the other initial guys for the amazing work
you did for OpenWapp and CoSeMe.

If you need to use WhatsApp in Firefox OS, you should move to LoquiIM, which
has great developers and they have implemented a lot of things of the new
WhatsApp protocol (encryption!)

Thanks!